requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.16.3
py-slippi==1.2.0
py-ubjson==0.13.0
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
termcolor==1.1.0
